refactor: block settings and add block menu to own files
