+++ packages/react-tinacms-inline/src/blocks/index.ts
@@ -22,3 +22,5 @@ export * from './inline-block-field-controls'
 export * from './inline-block-field'
 export * from './block'
 export * from './inline-block-textarea'
+export * from './block-settings'
+export * from './add-block-menu'
