+++ packages/react-tinacms-inline/src/blocks/inline-block-field-controls.tsx
@@ -18,39 +18,13 @@ limitations under the License.
 
 import * as React from 'react'
 import styled, { css } from 'styled-components'
-
-import {
-  FieldsBuilder,
-  BlockTemplate,
-  Modal,
-  ModalPopup,
-  ModalHeader,
-  ModalBody,
-  ModalActions,
-} from 'tinacms'
 import { useInlineBlock, useInlineBlocks } from './inline-field-blocks'
 import { useInlineForm } from '../inline-form'
-import {
-  radius,
-  color,
-  Button,
-  IconButton,
-  shadow,
-  font,
-} from '@tinacms/styles'
-import {
-  AddIcon,
-  ChevronUpIcon,
-  ChevronDownIcon,
-  TrashIcon,
-  SettingsIcon,
-} from '@tinacms/icons'
-import { useRef, useEffect } from 'react'
+import { AddBlockMenu } from './add-block-menu'
+import { BlockSettings } from './block-settings'
+import { radius, color, Button, IconButton } from '@tinacms/styles'
+import { ChevronUpIcon, ChevronDownIcon, TrashIcon } from '@tinacms/icons'
 
-/**
- *
- * TODO: I think this should be defined in `tinacms` and not with the rest of the inline stuff?
- */
 export interface BlocksControlsProps {
   children: any
   index: number
@@ -70,12 +44,12 @@ export function BlocksControls({ children, index }: BlocksControlsProps) {
   const { template } = useInlineBlock()
   const isFirst = index === 0
   const isLast = index === count - 1
-  const blockRef = useRef() as React.MutableRefObject<HTMLDivElement>
-  const blockMenuRef = useRef() as React.MutableRefObject<HTMLDivElement>
+  const blockRef = React.useRef() as React.MutableRefObject<HTMLDivElement>
+  const blockMenuRef = React.useRef() as React.MutableRefObject<HTMLDivElement>
 
   const removeBlock = () => remove(index)
 
-  useEffect(() => {
+  React.useEffect(() => {
     if (blockRef.current) {
       document.addEventListener('mousedown', clearActiveBlock)
     }
@@ -201,175 +175,4 @@ const BlockWrapper = styled.div<BlockWrapperProps>`
         opacity: 1 !important;
       }
     `};
-`
-
-interface AddBlockMenuProps {
-  addBlock(data: any): void
-  templates: BlockTemplate[]
-}
-
-export function AddBlockMenu({ templates, addBlock }: AddBlockMenuProps) {
-  const [isOpen, setOpen] = React.useState(false)
-
-  const handleOpenMenu = (event: React.MouseEvent) => {
-    event.preventDefault()
-    setOpen(isOpen => !isOpen)
-  }
-
-  React.useEffect(() => {
-    const setInactive = () => setOpen(false)
-    document.addEventListener('mouseup', setInactive, false)
-    return () => document.removeEventListener('mouseup', setInactive)
-  }, [])
-
-  templates = templates || []
-
-  return (
-    <>
-      <AddBlockButton onClick={handleOpenMenu} isOpen={isOpen} primary>
-        <AddIcon /> Add Block
-      </AddBlockButton>
-      <BlocksMenu isOpen={isOpen}>
-        {templates.map((template: BlockTemplate) => (
-          <BlockOption
-            key={template.label}
-            onClick={() => {
-              addBlock({
-                _template: template.type,
-                ...template.defaultItem,
-              })
-            }}
-          >
-            {template.label}
-          </BlockOption>
-        ))}
-      </BlocksMenu>
-    </>
-  )
-}
-
-interface BlockSettingsProps {
-  template: BlockTemplate
-}
-
-function BlockSettings({ template }: BlockSettingsProps) {
-  const [open, setOpen] = React.useState(false)
-  const noExtraFields = !(template.fields && template.fields.length)
-
-  if (noExtraFields) {
-    return null
-  }
-  return (
-    <>
-      <IconButton primary onClick={() => setOpen(p => !p)}>
-        <SettingsIcon />
-      </IconButton>
-      {open && (
-        <BlockSettingsModal template={template} close={() => setOpen(false)} />
-      )}
-    </>
-  )
-}
-
-function BlockSettingsModal({ template, close }: any) {
-  const { form } = useInlineForm()
-  const { name: blockName } = useInlineBlock()
-
-  const fields = template.fields.map((field: any) => ({
-    ...field,
-    name: `${blockName}.${field.name}`,
-  }))
-
-  return (
-    <Modal>
-      <ModalPopup>
-        <ModalHeader close={close}>Settings</ModalHeader>
-        <ModalBody>
-          <FieldsBuilder form={form} fields={fields} />
-        </ModalBody>
-        <ModalActions>
-          <Button onClick={close}>Cancel</Button>
-        </ModalActions>
-      </ModalPopup>
-    </Modal>
-  )
-}
-
-interface AddMenuProps {
-  isOpen?: boolean
-  active?: boolean
-}
-
-const AddBlockButton = styled(Button)<AddMenuProps>`
-  font-family: 'Inter', sans-serif;
-  display: flex;
-  align-items: center;
-  margin: 0 auto;
-
-  &:focus {
-    outline: none !important;
-  }
-
-  svg {
-    height: 70%;
-    width: auto;
-    margin-right: 0.5em;
-    transition: all 150ms ease-out;
-  }
-
-  ${props =>
-    props.open &&
-    css`
-      svg {
-        transform: rotate(45deg);
-      }
-    `};
-`
-
-const BlocksMenu = styled.div<AddMenuProps>`
-  min-width: 192px;
-  border-radius: ${radius()};
-  border: 1px solid ${color.grey(2)};
-  display: block;
-  position: absolute;
-  top: 0;
-  left: 50%;
-  transform: translate3d(-50%, 0, 0) scale3d(0.5, 0.5, 1);
-  opacity: 0;
-  pointer-events: none;
-  transition: all 150ms ease-out;
-  transform-origin: 50% 0;
-  box-shadow: ${shadow('big')};
-  background-color: white;
-  overflow: hidden;
-  ${props =>
-    props.isOpen &&
-    css`
-      opacity: 1;
-      pointer-events: all;
-      transform: translate3d(-50%, 48px, 0) scale3d(1, 1, 1);
-    `};
-`
-
-const BlockOption = styled.button`
-  font-family: 'Inter', sans-serif;
-  position: relative;
-  text-align: center;
-  font-size: ${font.size(0)};
-  padding: 0 12px;
-  height: 40px;
-  font-weight: 500;
-  width: 100%;
-  background: none;
-  cursor: pointer;
-  outline: none;
-  border: 0;
-  transition: all 85ms ease-out;
-  &:hover {
-    color: ${color.primary()};
-    background-color: #f6f6f9;
-  }
-  &:not(:last-child) {
-    border-bottom: 1px solid #efefef;
-  }
 `
